Move DefinitionResult back to Postgame directory.
