--- src/components/pages/GameContainer/Game/DefinitionResult/index.ts
@@ -1,1 +1,0 @@
-export { default as DefinitionResult } from './DefinitionResult';
